requirements.txt
psycopg2-binary==2.9.2
asyncpg==0.24.0
fastapi==0.75
fastapi-mail==1.0.8
databases[asyncpg]==0.5.5
typer==0.4.1
pydantic[dotenv]==1.8.2
pytest==6.2
asyncio==3.4.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
uvloop==0.16.0
uvicorn==0.15
gunicorn==20.1.0
requests==2.27.1
